Made the project a .exe, started working on comparators

+++ src/main/java/org/example/Models/TimeLog.java
@@ -53,7 +53,7 @@
             this.employeeId = employeeId;
         }
 
-        public int getHoursLogged() {
+        public Double getHoursLogged() {
             return hoursLogged;
         }
